refactor: fold snapshot into scraper, add GhScraper per-tick handle
Continues the GhClient direction (#348) by collapsing the plugin → scraper
→ snapshot → fetch stack one more layer.

- Fold snapshot.ts into scraper.ts. `snapshotPr`/`snapshotIssue` +
  `stripInternals` + `sortedIds`/`indexById` become module-private helpers;
  only `PrSnapInternal`/`IssueSnapInternal` stay exported (diff.ts still
  consumes them).
- Introduce `GhScraper`, a per-tick handle bundling `GhClient + agentLogins`.
  `scrapePr(repo, number, prevSnap)` / `scrapeIssue(repo, number, prevIssue)`
  as methods. Plugins construct one in `runTick` instead of threading
  `client` and `agentLogins` through every scrape callsite.

Plugin runTick callsites drop two args per scrape. `agentLogins` no longer
rides 4 hops to reach `formatCommentEvent` — it lives on the scraper and is
applied where event computation happens.

Tests: `plugin.test.ts` spies switch from `spyOn(scraper, 'scrapePr')` to
`spyOn(GhScraper.prototype, 'scrapePr')` (same pattern as the existing
`GhPluginBase.prototype.observeRateLimit` spy on line 381). `scraper.test.ts`
and `diff.test.ts` retarget snapshot type imports to scraper.ts.

No behavior change. 510 tests pass, typecheck + lint clean.

## Human-Claude Interaction Log

### Human prompts (VERBATIM - include typos, informal language, COMPLETE text):

1. "/worker roost 351 AvesAlight/roost feat/351-restructure-plugins alex" (from alex)
   → Claude: read issue #351 + PR #348 context, explored the github plugin
     stack (base/github-api/snapshot/scraper/prs-plugin/issues-plugin),
     posted a plan to #roost-issue-351 proposing (1) fold snapshot.ts into
     scraper.ts and (2) introduce a GhScraper class bundling client +
     agentLogins, with a stretch option to extract the runTick template.

2. "plan looks right. fold + GhScraper handle, skip the stretch (you called
    it — divergence is real). proceed." (from roost-lead-pm in IRC)
   → Claude: executed the two-move plan: rewrote scraper.ts with private
     snapshot helpers + GhScraper class, deleted snapshot.ts, retargeted
     imports in diff.ts + two test files, updated prs-plugin/issues-plugin
     to construct a GhScraper, switched plugin.test.ts spies to
     GhScraper.prototype. Typecheck/lint/tests all green.
