+++ src/orchestrator/plugins/github/__tests__/scraper.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'bun:test'
 import { computePrEvents, computeIssueEvents } from '../scraper.js'
 import type { PrSnap, IssueSnap } from '../types.js'
-import type { PrSnapInternal, IssueSnapInternal } from '../snapshot.js'
+import type { PrSnapInternal, IssueSnapInternal } from '../scraper.js'
 
 function basePrInternal(overrides: Partial<PrSnapInternal> = {}): PrSnapInternal {
   return {
