+++ src/orchestrator/plugins/github/github-api.ts
@@ -280,8 +280,8 @@ export function labelNames(labels: GhLabel[] | null | undefined): string[] {
 // GhClient — per-plugin handle that owns the PluginLogger and exposes the
 // fetch surface as instance methods. Every gh call lands in spawnGh via the
 // private api()/graphql() shape helpers, so retry stays the universal contract.
-// Plugins construct one at boot (see GhPluginBase) and pass it down through
-// snapshot/scraper instead of threading `log` through each callsite.
+// Plugins construct one at boot (see GhPluginBase) and hand it to a per-tick
+// GhScraper instead of threading `log` through each callsite.
 export class GhClient {
   constructor(private readonly log: PluginLogger) {}
