+++ src/orchestrator/plugins/github/scraper.ts
@@ -1,16 +1,113 @@
-// Scraper helpers — one per GH entity type. Each fetches the current snapshot
-// and computes what events to emit vs. the previous state.
+// Scraper — fetches a PR/issue snapshot and diffs it against the previous one
+// to produce events. Snapshot construction (fetch + index for diff lookups)
+// is module-private here; the only public surfaces are `GhScraper` (the
+// per-tick handle plugins use), `computePrEvents`/`computeIssueEvents` (pure
+// event computation tested directly), and the `*Internal` snapshot types that
+// `diff.ts` consumes.
 //
 // prevSnap meanings:
 //   undefined → seeding tick (global --seed or no prior state); emit nothing
 //   null      → entity is new to the watch list; emit seed/backlog events
 //   PrSnap    → normal tick; diff against prev and emit change events
 import type { PrSnap, IssueSnap } from './types.js'
-import type { GhClient } from './github-api.js'
-import { snapshotPr, snapshotIssue, stripInternals } from './snapshot.js'
-import type { PrSnapInternal, IssueSnapInternal } from './snapshot.js'
-import { diffPr, diffIssue } from './diff.js'
-import type { OrchestratorEvent } from './diff.js'
+import type { GhClient, GhComment, GhReview } from './github-api.js'
+import { labelNames } from './github-api.js'
+import { diffPr, diffIssue, type OrchestratorEvent } from './diff.js'
+
+// ---- Snapshot types & helpers (module-private outside the *Internal types,
+// which diff.ts imports) ----------------------------------------------------
+
+export interface PrSnapInternal extends PrSnap {
+  _review_comments_by_id: Record<number, GhComment>
+  _conversation_comments_by_id: Record<number, GhComment>
+  _reviews_by_id: Record<number, GhReview>
+}
+
+export interface IssueSnapInternal extends IssueSnap {
+  _comments_by_id: Record<number, GhComment>
+}
+
+function stripInternals(snap: PrSnapInternal): PrSnap
+function stripInternals(snap: IssueSnapInternal): IssueSnap
+function stripInternals(snap: PrSnapInternal | IssueSnapInternal): PrSnap | IssueSnap {
+  const out: Record<string, unknown> = {}
+  for (const [k, v] of Object.entries(snap)) {
+    if (!k.startsWith('_')) out[k] = v
+  }
+  return out as unknown as PrSnap | IssueSnap
+}
+
+function sortedIds<T extends { id?: number }>(items: T[]): number[] {
+  return items.filter(c => c.id != null).map(c => c.id as number).sort((a, b) => a - b)
+}
+
+function indexById<T extends { id?: number }>(items: T[]): Record<number, T> {
+  const out: Record<number, T> = {}
+  for (const item of items) {
+    if (item.id != null) out[item.id] = item
+  }
+  return out
+}
+
+async function snapshotPr(
+  client: GhClient,
+  repo: string,
+  number: number,
+  prevSnap?: PrSnap | null
+): Promise<PrSnapInternal> {
+  const view = await client.fetchPr(repo, number)
+  const [reviewComments, convComments, reviews] = await Promise.all([
+    client.fetchPrReviewComments(repo, number),
+    client.fetchPrConversationComments(repo, number),
+    client.fetchPrReviews(repo, number),
+  ])
+
+  const curHead = view.head_oid
+  const linkedIssues =
+    prevSnap && prevSnap.head_oid === curHead
+      ? prevSnap.linked_issues ?? []
+      : await client.fetchPrLinkedIssues(repo, number)
+
+  return {
+    repo,
+    number,
+    title: view.title,
+    url: view.url,
+    head_ref: view.head_ref,
+    head_oid: curHead,
+    is_draft: view.is_draft,
+    merged: view.merged_at != null,
+    state: view.state,
+    labels: labelNames(view.labels),
+    ci_state: view.ci_state,
+    linked_issues: linkedIssues,
+    seen_review_comment_ids: sortedIds(reviewComments),
+    seen_conversation_comment_ids: sortedIds(convComments),
+    seen_review_ids: sortedIds(reviews),
+    _review_comments_by_id: indexById(reviewComments),
+    _conversation_comments_by_id: indexById(convComments),
+    _reviews_by_id: indexById(reviews),
+  }
+}
+
+async function snapshotIssue(client: GhClient, repo: string, number: number): Promise<IssueSnapInternal> {
+  const [issue, comments] = await Promise.all([
+    client.fetchIssue(repo, number),
+    client.fetchIssueComments(repo, number),
+  ])
+  return {
+    repo,
+    number,
+    title: issue.title ?? null,
+    url: issue.html_url ?? null,
+    state: issue.state ?? null,
+    labels: labelNames(issue.labels),
+    seen_comment_ids: sortedIds(comments),
+    _comments_by_id: indexById(comments),
+  }
+}
+
+// ---- Event computation (pure; tested directly) ----------------------------
 
 export interface ScrapeResult<T> {
   snap: T
@@ -63,27 +160,37 @@ export function computeIssueEvents(
   return events
 }
 
-export async function scrapePr(
-  client: GhClient,
-  repo: string,
-  number: number,
-  prevSnap: PrSnap | null | undefined,
-  agentLogins: Set<string>
-): Promise<ScrapeResult<PrSnap>> {
-  const snap = await snapshotPr(client, repo, number, prevSnap ?? undefined)
-  const { events, nextWarnedNoLinked } = computePrEvents(snap, prevSnap, agentLogins)
-  const stripped = stripInternals(snap) as PrSnap
-  stripped.warned_no_linked = nextWarnedNoLinked
-  return { snap: stripped, events }
-}
+// ---- GhScraper — per-tick handle bundling client + agentLogins ------------
+//
+// Plugins construct one per `runTick` (agentLogins can drift between ticks
+// when an operator edits config), then call `scrapePr`/`scrapeIssue` per
+// watched entry. Replaces threading `client` and `agentLogins` through every
+// scrape callsite.
 
-export async function scrapeIssue(
-  client: GhClient,
-  repo: string,
-  number: number,
-  prevIssue: IssueSnap | null | undefined,
-  agentLogins: Set<string>
-): Promise<ScrapeResult<IssueSnap>> {
-  const snap = await snapshotIssue(client, repo, number)
-  return { snap: stripInternals(snap) as IssueSnap, events: computeIssueEvents(snap, prevIssue, agentLogins) }
+export class GhScraper {
+  constructor(
+    private readonly client: GhClient,
+    private readonly agentLogins: Set<string>,
+  ) {}
+
+  async scrapePr(
+    repo: string,
+    number: number,
+    prevSnap: PrSnap | null | undefined,
+  ): Promise<ScrapeResult<PrSnap>> {
+    const snap = await snapshotPr(this.client, repo, number, prevSnap ?? undefined)
+    const { events, nextWarnedNoLinked } = computePrEvents(snap, prevSnap, this.agentLogins)
+    const stripped = stripInternals(snap) as PrSnap
+    stripped.warned_no_linked = nextWarnedNoLinked
+    return { snap: stripped, events }
+  }
+
+  async scrapeIssue(
+    repo: string,
+    number: number,
+    prevIssue: IssueSnap | null | undefined,
+  ): Promise<ScrapeResult<IssueSnap>> {
+    const snap = await snapshotIssue(this.client, repo, number)
+    return { snap: stripInternals(snap) as IssueSnap, events: computeIssueEvents(snap, prevIssue, this.agentLogins) }
+  }
 }
